progress 15

+++ easy/70.cpp
@@ -13,7 +13,7 @@ Each time you can either climb 1 or 2 steps. In how many distinct ways can you c
 
 int climbStairs(int n) {
   std::vector<int> ways = {1, 1};
-  for (int i = 2; i < n + 1; ++i) {
+  for (size_t i = 2; i < n + 1; ++i) {
     ways.push_back(ways[i - 1] + ways[i - 2]);
   }
   return ways[n];
